backend/requirements.txt
# FastAPI & ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# MongoDB + BSON
pymongo==4.6.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Data Validation
pydantic==2.5.0
